project/requirements.txt
fastapi==0.65.2
uvicorn==0.14
pydantic==1.6.2
SQLAlchemy
alembic==1.4.2
FastAPI-SQLAlchemy==0.1.6
psycopg2-binary==2.8.6
sqlalchemy_utils
environ-config==21.2.0
requests==2.24.0